==> 1_attributes_literature.Rmd <==
---
title: "Summary of attributes from the literature review"
author: "Adrian Barnett"
date: "2023-01-31"
output: word_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400)
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation
library(ggplot2)
library(dplyr)
library(flextable)
library(stringr)

# get data
load('data/0_literature.RData') # from 0_read_google_sheet.R
n_studies = length(unique(long$author))
n_attributes = length(unique(long$attribute))
```

There were `r n_studies` studies included in this stage of the literature review. They examined `r n_attributes` attributes.

# Frequencies by category

The table below shows the number of attributes mentioned in each category.

```{r}
tab = group_by(long, group) %>%
  summarise(n = sum(value)) %>%
  ungroup() %>%
  arrange(-n)
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

# Frequencies by attribute

The table below shows the top ten attributes.

```{r}
stats = group_by(long, attribute) %>%
  summarise(n = sum(value)) %>%
  ungroup() 
tab = arrange(stats, -n) %>%
  slice(1:10)
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
#
n_zero = filter(stats, n==0) %>% nrow()
```


There were `r n_zero` attributes that were never mentioned, these were:

```{r, results='asis'}
zero = filter(stats, n==0)
for (k in 1:nrow(zero)){
  cat('* ', zero$attribute[k], '\n', sep='')
}
```

# Frequencies by paper

```{r}
stats = group_by(long, author) %>%
  summarise(n = sum(value)) %>%
  ungroup() 
tab = summarise(stats,
                min = min(n),
                q1 = round(quantile(n,0.25)),
            median = round(median(n)),
            q3 = round(quantile(n,0.75)),
            max = max(n))
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
#
n_zero = filter(stats, n==0) %>% nrow()
```

The table shows summary statistics on the number of attributes mentioned per paper.
There were `r n_zero` papers with no attributes.

# Tile plot

Below is a tile plot showing the attributes mentioned by each paper. The results are grouped by the six categories.

```{r, fig.width=9, fig.height=11}
#
long = mutate(long, 
              attribute = str_replace(attribute, "^\\w{1}", toupper), # first word should be capital
              author = str_replace_all(author, '_', ' '))
# for ordering
levels = c('Journal','Impact','Paper characteristics','Authors','Other','Activity')
long = mutate(long, group = factor(group, levels=levels))
#
tile = ggplot(data = long, aes(y=attribute, x=author, fill=value))+ #, alpha = group))+ alpha makes shading per group
  geom_tile()+
  scale_fill_manual(NULL, values=c('beige','dark red'))+
  theme_minimal()+
  theme(
    strip.text = element_text(size = 7),
    panel.spacing = unit(1, "lines"), # gaps between panels
    plot.margin = margin(t = 0, r = 0, b = 0, l = 0, unit = "pt"),
    legend.position = 'none',
        panel.grid.minor = element_blank())+
  scale_x_discrete(labels=NULL) + # No point showing papers, too small
  facet_grid(rows = vars(group), scales = "free", space = "free") +
  xlab('')+
  ylab('')
tile
ggsave(filename = 'figures/literature_review.jpg', device = 'jpeg', plot = tile, units = 'in', width=9, height=12, dpi=600)
#ggsave(filename = 'figures/literature_review.eps', device = 'eps', plot = tile, width=9, height=12) # for protocol
```
